Move container functions to container module
And move image functions to image module
