+++ aws_mp_utils/changeset.py
@@ -21,14 +21,11 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import json
 import re
 import time
 
 import boto3
 import botocore.exceptions as boto_exceptions
-
-import jmespath
 
 from aws_mp_utils.exceptions import AWSMPUtilsException
 
@@ -58,86 +55,6 @@ def get_change_set_status(
         # 'Status':'PREPARING'|'APPLYING'|'SUCCEEDED'|'CANCELLED'|'FAILED'
         status = response['Status'].lower()
         return status
-
-
-def create_restrict_version_change_doc(
-    entity_id: str,
-    delivery_option_id: str
-) -> dict:
-    """
-    Creates a restrict delivery option request dictionary.
-    For use with submitting a changeset to delete an image
-    or container version from a product.
-    """
-    data = {
-        'ChangeType': 'RestrictDeliveryOptions',
-        'Entity': {
-            'Type': 'AmiProduct@1.0',
-            'Identifier': entity_id
-        }
-    }
-    details = {
-        'DeliveryOptionIds': [delivery_option_id]
-    }
-
-    data['Details'] = json.dumps(details)
-    return data
-
-
-def create_add_version_change_doc(
-    entity_id: str,
-    version_title: str,
-    ami_id: str,
-    access_role_arn: str,
-    release_notes: str,
-    os_name: str,
-    os_version: str,
-    usage_instructions: str,
-    recommended_instance_type: str,
-    ssh_user: str = 'ec2-user',
-    ingress_rules: list = None,
-) -> dict:
-    if not ingress_rules:
-        ingress_rules = [{
-            'FromPort': 22,
-            'IpProtocol': 'tcp',
-            'IpRanges': ['0.0.0.0/0'],
-            'ToPort': 22
-        }]
-
-    data = {
-        'ChangeType': 'AddDeliveryOptions',
-        'Entity': {
-            'Type': 'AmiProduct@1.0',
-            'Identifier': entity_id
-        }
-    }
-
-    details = {
-        'Version': {
-            'VersionTitle': version_title,
-            'ReleaseNotes': release_notes
-        },
-        'DeliveryOptions': [{
-            'Details': {
-                'AmiDeliveryOptionDetails': {
-                    'UsageInstructions': usage_instructions,
-                    'RecommendedInstanceType': recommended_instance_type,
-                    'AmiSource': {
-                        'AmiId': ami_id,
-                        'AccessRoleArn': access_role_arn,
-                        'UserName': ssh_user,
-                        'OperatingSystemName': os_name,
-                        'OperatingSystemVersion': os_version
-                    },
-                    'SecurityGroups': ingress_rules
-                }
-            }
-        }]
-    }
-
-    data['Details'] = json.dumps(details)
-    return data
 
 
 def start_mp_change_set(
@@ -209,219 +126,3 @@ def get_ongoing_change_id_from_error(message: str) -> str:
         raise AWSMPUtilsException(
             f'Unable to extract changeset id from aws err response: {message}'
         )
-
-
-def get_image_delivery_option_id(
-    client: boto3.client,
-    entity_id: str,
-    ami_id: str
-) -> str:
-    """
-    Return delivery option id for image matching ami id in given offer
-    """
-    entity = client.describe_entity(
-        Catalog='AWSMarketplace',
-        EntityId=entity_id
-    )
-
-    """
-    Example describe entity output:
-    {
-        "Details": {
-            "Versions": [
-                {
-                    "Sources": [
-                        {
-                            "Image": "ami-123",
-                            "Id": "1234"
-                        }
-                    ],
-                    "DeliveryOptions": [
-                        {
-                            "Id": "4321",
-                            "SourceId": "1234"
-                        }
-                    ]
-                }
-            ]
-        }
-    }
-    """
-    details = entity['DetailsDocument']
-
-    source_id = jmespath.search(
-        f"Versions[?Sources[?Image=='{ami_id}']] | [0].Sources | [0].Id",
-        details
-    )
-
-    if not source_id:
-        return None
-
-    delivery_option_id = jmespath.search(
-        f"Versions[?DeliveryOptions[?SourceId=='{source_id}']] | "
-        "[0].DeliveryOptions | [0].Id",
-        details
-    )
-
-    return delivery_option_id
-
-
-def get_helm_delivery_option_id(
-    client: boto3.client,
-    entity_id: str,
-    version_title: str
-) -> str:
-    """
-    Provides the id of the delivery option to be updated. The id of
-    the delivery option is selected from the version of the product
-    that matches the version_title provided. The delivery option is
-    the first delivery option of the type 'Helm' in that version.
-    Example describe entity output:
-    {
-        "Details": {
-            "Versions": [
-                {
-                    "VersionTitle": "Product 1.2.3",
-                    "DeliveryOptions": [
-                        {
-                            "Id": "4321",
-                            "Type": "Helm"
-                        }
-                    ]
-                }
-            ]
-        }
-    }
-    """
-    entity = client.describe_entity(
-        Catalog='AWSMarketplace',
-        EntityId=entity_id
-    )
-    details = entity['DetailsDocument']
-
-    result = jmespath.search(
-        f"Versions[?VersionTitle=='{version_title}'] | [0]"
-        ".DeliveryOptions[?Type=='Helm'] | [0].Id",
-        details
-    )
-    return result
-
-
-def gen_add_delivery_options_changeset(
-    entity_id: str,
-    version_title: str,
-    release_notes: str,
-    delivery_option_title: str,
-    compatible_services: list,
-    container_images: list,
-    helm_chart_uri: str,
-    helm_chart_description: str,
-    usage_instructions: str,
-    quick_launch_enabled: bool,
-    marketplace_service_account_name: str,
-    release_name: str,
-    namespace: str,
-    override_parameters: list
-) -> dict:
-    """
-    Function to generate a marketplace changeset of AddDeliveryOptions type
-    https://docs.aws.amazon.com/marketplace-catalog/latest/api-reference/
-    container-products.html#working-with-container-products
-    """
-
-    data = {
-        'ChangeType': 'AddDeliveryOptions',
-        'Entity': {
-            'Type': 'ContainerProduct@1.0',
-            'Identifier': entity_id
-        }
-    }
-
-    details = {
-        'Version': {
-            'VersionTitle': version_title,
-            'ReleaseNotes': release_notes
-        },
-        'DeliveryOptions': [{
-            'DeliveryOptionTitle': delivery_option_title,
-            'Details': {
-                'HelmDeliveryOptionDetails': {
-                    'CompatibleServices': compatible_services,
-                    'ContainerImages': container_images,
-                    'HelmChartUri': helm_chart_uri,
-                    'Description': helm_chart_description,
-                    'UsageInstructions': usage_instructions,
-                    'QuickLaunchEnabled': quick_launch_enabled,
-                    'MarketplaceServiceAccountName':
-                        marketplace_service_account_name,
-                    'ReleaseName': release_name,
-                    'Namespace': namespace,
-                    'OverrideParameters': override_parameters
-                }
-            }
-        }]
-    }
-
-    data['Details'] = json.dumps(details)
-    return data
-
-
-def gen_update_delivery_options_changeset(
-    entity_id: str,
-    version_title: str,
-    release_notes: str,
-    delivery_option_title: str,
-    compatible_services: list,
-    container_images: list,
-    helm_chart_uri: str,
-    helm_chart_description: str,
-    usage_instructions: str,
-    quick_launch_enabled: bool,
-    marketplace_service_account_name: str,
-    release_name: str,
-    namespace: str,
-    override_parameters: list,
-    delivery_option_id: str
-) -> dict:
-    """
-    Function to generate a marketplace changeset of UpdateDeliveryOptions
-    type https://docs.aws.amazon.com/marketplace-catalog/latest/
-    api-reference/container-products.html#working-with-container-products
-    """
-
-    data = {
-        'ChangeType': 'UpdateDeliveryOptions',
-        'Entity': {
-            'Type': 'ContainerProduct@1.0',
-            'Identifier': entity_id
-        }
-    }
-
-    details = {
-        'Version': {
-            'VersionTitle': version_title,
-            'ReleaseNotes': release_notes
-        },
-        'DeliveryOptions': [{
-            'Id': delivery_option_id,
-            'Details': {
-                'HelmDeliveryOptionDetails': {
-                    'DeliveryOptionTitle': delivery_option_title,
-                    'CompatibleServices': compatible_services,
-                    'ContainerImages': container_images,
-                    'HelmChartUri': helm_chart_uri,
-                    'Description': helm_chart_description,
-                    'UsageInstructions': usage_instructions,
-                    'QuickLaunchEnabled': quick_launch_enabled,
-                    'MarketplaceServiceAccountName':
-                        marketplace_service_account_name,
-                    'ReleaseName': release_name,
-                    'Namespace': namespace,
-                    'OverrideParameters': override_parameters
-                }
-            }
-        }]
-    }
-
-    data['Details'] = json.dumps(details)
-    return data
