+++ tests/test_changeset.py
@@ -6,16 +6,11 @@ import pytest
 
 from unittest.mock import call, Mock
 
+from aws_mp_utils.image import create_add_version_change_doc
 from aws_mp_utils.changeset import (
     get_change_set,
     get_change_set_status,
-    start_mp_change_set,
-    create_add_version_change_doc,
-    create_restrict_version_change_doc,
-    get_image_delivery_option_id,
-    get_helm_delivery_option_id,
-    gen_add_delivery_options_changeset,
-    gen_update_delivery_options_changeset
+    start_mp_change_set
 )
 
 # botocore ClientExceptionFactory
@@ -270,170 +265,3 @@ def test_start_mp_change_set_ongoing_change_GenericBotoException():
             call(**start_changeset_params),
         ],
     )
-
-
-def test_create_restrict_version_change_doc():
-    expected = {
-        'ChangeType': 'RestrictDeliveryOptions',
-        'Entity': {
-            'Type': 'AmiProduct@1.0',
-            'Identifier': '123456789'
-        }
-    }
-    details = {
-        'DeliveryOptionIds': ['987654321']
-    }
-    expected['Details'] = json.dumps(details)
-
-    actual = create_restrict_version_change_doc('123456789', '987654321')
-    assert expected == actual
-
-
-def test_get_image_delivery_option_id():
-    details = {
-        "Versions": [
-            {
-                "Sources": [
-                    {
-                        "Image": "ami-123",
-                        "Id": "1234"
-                    }
-                ],
-                "DeliveryOptions": [
-                    {
-                        "Id": "4321",
-                        "SourceId": "1234"
-                    }
-                ]
-            }
-        ]
-    }
-
-    entity = {
-        'DetailsDocument': details
-    }
-    client = Mock()
-    client.describe_entity.return_value = entity
-
-    did = get_image_delivery_option_id(
-        client,
-        '1234589',
-        'ami-123',
-    )
-    assert did == '4321'
-
-    # Test no image match found
-    details['Versions'][0]['Sources'][0]['Image'] = 'ami-321'
-    entity['Details'] = json.dumps(details)
-
-    did = get_image_delivery_option_id(
-        client,
-        '1234589',
-        'ami-123',
-    )
-    assert did is None
-
-
-def test_get_helm_delivery_option_id():
-    details = {
-        "Versions": [
-            {
-                "VersionTitle": "Product 1.2.3",
-                "DeliveryOptions": [
-                    {
-                        "Id": "4321",
-                        "Type": "Helm"
-                    }
-                ]
-            }
-        ]
-    }
-
-    entity = {
-        'DetailsDocument': details
-    }
-    client = Mock()
-    client.describe_entity.return_value = entity
-
-    did = get_helm_delivery_option_id(
-        client,
-        '1234589',
-        'Product 1.2.3',
-    )
-    assert did == '4321'
-
-    # Test no image match found
-    details['Versions'][0]['VersionTitle'] = 'Product 1.2.4'
-    entity['Details'] = json.dumps(details)
-
-    did = get_helm_delivery_option_id(
-        client,
-        '1234589',
-        'Product 1.2.3',
-    )
-    assert did is None
-
-
-def test_gen_add_delivery_options_changeset():
-    response = gen_add_delivery_options_changeset(
-        '123',
-        'Chart 1.2.3',
-        'Release Notes...',
-        'Chart - 1.2.3',
-        ['EKS'],
-        ['123.dkr.ecr.us-east-1.amazonaws.com/sellername/reponame1:1'],
-        '123.dkr.ecr.us-east-1.amazonaws.com/sellername/reponame1:helmchart1',
-        'Helm chart description',
-        'Usage instructions',
-        True,
-        'Service account name',
-        'Optional release name',
-        'Optional k8s namespace',
-        [
-            {
-                "Key": "HelmKeyName1",
-                "DefaultValue": "${AWSMP_LICENSE_SECRET}",
-                "Metadata":
-                {
-                    "Label": "AWS CloudFormation template field label",
-                    "Description": "AWS CloudFormation field description",
-                    "Obfuscate": False
-                }
-            }
-        ]
-    )
-    assert 'Details' in response
-    assert response['ChangeType'] == 'AddDeliveryOptions'
-
-
-def test_gen_update_delivery_options_changeset():
-    response = gen_update_delivery_options_changeset(
-        '123',
-        'Chart 1.2.3',
-        'Release Notes...',
-        'Chart - 1.2.3',
-        ['EKS'],
-        ['123.dkr.ecr.us-east-1.amazonaws.com/sellername/reponame1:1'],
-        '123.dkr.ecr.us-east-1.amazonaws.com/sellername/reponame1:helmchart1',
-        'Helm chart description',
-        'Usage instructions',
-        True,
-        'Service account name',
-        'Optional release name',
-        'Optional k8s namespace',
-        [
-            {
-                "Key": "HelmKeyName1",
-                "DefaultValue": "${AWSMP_LICENSE_SECRET}",
-                "Metadata":
-                {
-                    "Label": "AWS CloudFormation template field label",
-                    "Description": "AWS CloudFormation field description",
-                    "Obfuscate": False
-                }
-            }
-        ],
-        '123-321'
-    )
-    assert 'Details' in response
-    assert response['ChangeType'] == 'UpdateDeliveryOptions'
